split SD structs into common and SD-specific modules
